Update MinimalEventCounterSource.cs
StopWatch.TotalMiliseconds outputs a long.
Also WriteEvent has a specific overload for int, string, long. which is significantly faster than the int, Object[] args one that was used in this snippet.

+++ docs/core/diagnostics/snippets/EventCounters/MinimalEventCounterSource.cs
@@ -14,7 +14,7 @@ public sealed class MinimalEventCounterSource : EventSource
             DisplayUnits = "ms"
         };
 
-    public void Request(string url, float elapsedMilliseconds)
+    public void Request(string url, long elapsedMilliseconds)
     {
         WriteEvent(1, url, elapsedMilliseconds);
         _requestCounter?.WriteMetric(elapsedMilliseconds);
